refactoring of utils files
