+++ sktime/utils/tests/test_utils_transformations.py
@@ -3,8 +3,8 @@ import pandas as pd
 import pytest
 
 from sktime.utils.testing import generate_df_from_array, generate_polynomial_series
-from sktime.utils.transformations import fit_trend, remove_trend, add_trend
-from sktime.utils.transformations import tabularize
+from sktime.utils.time_series import fit_trend, remove_trend, add_trend
+from sktime.utils.data_container import tabularize
 
 
 def test_tabularize():
